# backend/web/router.go
package web

import (
	"encoding/json"
	"errors"
	"net/http"
	"strconv"

	"pokemon-binder-finder/model"
	"pokemon-binder-finder/pokemontcg"
	"pokemon-binder-finder/service"
)

func NewRouter(catalog *pokemontcg.Catalog, searches *service.SearchService, assetsDir string) http.Handler {
	mux := http.NewServeMux()

	mux.HandleFunc("GET /api/health", func(w http.ResponseWriter, _ *http.Request) {
		writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
	})

	mux.HandleFunc("GET /api/cards", func(w http.ResponseWriter, r *http.Request) {
		limit, _ := strconv.Atoi(r.URL.Query().Get("limit"))
		if limit <= 0 || limit > 50 {
			limit = 20
		}
		cards, err := catalog.Search(r.Context(), r.URL.Query().Get("query"), limit)
		if err != nil {
			writeError(w, http.StatusInternalServerError, err)
			return
		}
		writeJSON(w, http.StatusOK, cards)
	})
	mux.HandleFunc("POST /api/catalog/sync", func(w http.ResponseWriter, r *http.Request) {
		if err := catalog.Sync(r.Context()); err != nil {
			writeError(w, http.StatusBadGateway, err)
			return
		}
		writeJSON(w, http.StatusOK, map[string]string{"status": "synced"})
	})
	mux.HandleFunc("POST /api/search-jobs", func(w http.ResponseWriter, r *http.Request) {
		var input struct {
			CardID       string `json:"cardId"`
			ListingQuery string `json:"listingQuery"`
		}
		if err := json.NewDecoder(r.Body).Decode(&input); err != nil {
			writeError(w, http.StatusBadRequest, err)
			return
		}
		job, err := searches.Create(r.Context(), input.CardID, input.ListingQuery)
		if err != nil {
			writeError(w, http.StatusBadRequest, err)
			return
		}
		writeJSON(w, http.StatusAccepted, job)
	})
	mux.HandleFunc("GET /api/search-jobs/{id}", func(w http.ResponseWriter, r *http.Request) {
		job, err := searches.Get(r.Context(), r.PathValue("id"))
		if err != nil {
			writeStoreError(w, err)
			return
		}
		writeJSON(w, http.StatusOK, job)
	})
	mux.HandleFunc("GET /api/search-jobs/{id}/results", func(w http.ResponseWriter, r *http.Request) {
		bucket := r.URL.Query().Get("bucket")
		if bucket != model.BucketConfirmed && bucket != model.BucketPossible {
			writeError(w, http.StatusBadRequest, errors.New("bucket must be confirmed or possible"))
			return
		}
		results, err := searches.Results(r.Context(), r.PathValue("id"), bucket)
		if err != nil {
			writeStoreError(w, err)
			return
		}
		writeJSON(w, http.StatusOK, results)
	})
	if assetsDir != "" {
		mux.Handle("/api/assets/", http.StripPrefix("/api/assets/", http.FileServer(http.Dir(assetsDir))))
	}
	return withCORS(mux)
}

func writeStoreError(w http.ResponseWriter, err error) {
	if errors.Is(err, model.ErrNotFound) {
		writeError(w, http.StatusNotFound, err)
		return
	}
	writeError(w, http.StatusInternalServerError, err)
}

func writeJSON(w http.ResponseWriter, status int, value any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(value)
}

func writeError(w http.ResponseWriter, status int, err error) {
	writeJSON(w, status, map[string]string{"error": err.Error()})
}

func withCORS(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Access-Control-Allow-Origin", "*")
		w.Header().Set("Access-Control-Allow-Headers", "Content-Type")
		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, OPTIONS")
		if r.Method == http.MethodOptions {
			w.WriteHeader(http.StatusNoContent)
			return
		}
		next.ServeHTTP(w, r)
	})
}